requirements.txt
pyyaml>=6.0,<7
aiohttp>=3.9,<4
python-telegram-bot>=21.0,<22
anthropic>=0.39,<1
openai>=1.50,<2
